capcomposer/requirements/base.txt
wagtail>=7.0
djangorestframework-xml
python-magic
shapely
wagtail-icon-chooser>=0.3.1
wagtail-cache==3.0.0
adm-boundary-manager>=0.2.5
wagtail-humanitarian-icons>=2.0.0
wagtail-modelchooser>=4.0.1
paho-mqtt>=2.1.0
xmltodict>=0.14.2
pdf2image>=1.17.0
wagtail-metadata>=5.0.0
lxml>=5.4.0
signxml>=4.0.3
loguru>=0.7.3
weasyprint>=65.1
requests>=2.32.3
pytz>=2025.2
wagtail-lazyimages==0.1.5
wagtail-font-awesome-svg>=1.0.1
capvalidator==0.1.0.dev4
feedparser>=6.0.11
staticmap==0.5.7
wagtail-newsletter==0.2.2
mailchimp-marketing==3.0.80
mrml>=0.2